Feat: Remove Restrictions On Animated Items
For anyone who's wondering about this, I have actually tested this within my portfolio website locally and it works, this is also my first pr to a package so hopefully someone understands. I may not have a lot of experience but I don't think there should really be a restriction on how many children should be within an element, that's just my honest opinion.